Move CoroutineUtilities to WTF https://bugs.webkit.org/show_bug.cgi?id=289520 rdar://146725359
Reviewed by Abrar Rahman Protyasha.

They are low-level utilities with no dependencies except C++20.
This allows us to share CoroutineHandle.

* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/CMakeLists.txt:
* Source/WTF/wtf/CoroutineUtilities.h: Renamed from Source/WebKit/Platform/CoroutineUtilities.h.
(WTF::CoroutineHandle::CoroutineHandle):
(WTF::CoroutineHandle::operator=):
(WTF::CoroutineHandle::~CoroutineHandle):
(WTF::CoroutineHandle::handle const):
(WTF::Task::promise_type::get_return_object):
(WTF::Task::promise_type::initial_suspend):
(WTF::Task::promise_type::unhandled_exception):
(WTF::Task::promise_type::return_void):
* Source/WTF/wtf/Forward.h:
* Source/WebKit/Platform/IPC/HandleMessage.h:
(IPC::callMemberFunctionCoroutine):
(IPC::callMemberFunctionCoroutineVoid):
* Source/WebKit/UIProcess/WebFullScreenManagerProxy.cpp:
* Source/WebKit/UIProcess/WebFullScreenManagerProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Tools/TestWebKitAPI/CoroutineUtilities.h:
(TestWebKitAPI::ConnectionTask::promise_type::get_return_object):
(TestWebKitAPI::CoroutineHandle::CoroutineHandle): Deleted.
(TestWebKitAPI::CoroutineHandle::~CoroutineHandle): Deleted.
(TestWebKitAPI::Task::promise_type::get_return_object): Deleted.
(TestWebKitAPI::Task::promise_type::initial_suspend): Deleted.
(TestWebKitAPI::Task::promise_type::unhandled_exception): Deleted.
(TestWebKitAPI::Task::promise_type::return_void): Deleted.
* Tools/TestWebKitAPI/Tests/WebKitCocoa/CookieAcceptPolicy.mm:
(TEST(WKHTTPCookieStore, CookiePolicy)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/LoadAlternateHTMLString.mm:
(TEST(WebKit, LoadMoreThan4GB)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/Navigation.mm:
(setupWebViewForPrivateTokenTests):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/NetworkProcess.mm:
(TEST(_WKDataTask, Basic)):
(TEST(WKWebView, CrossOriginDoubleRedirectAuthentication)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/ResourceLoadStatistics.mm:
(TEST(ResourceLoadStatistics, StorageAccessSupportMultipleSubFrameDomains)):
(TEST(ResourceLoadStatistics, StorageAccessGrantMultipleSubFrameDomains)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SOAuthorizationTests.mm:
(TestWebKitAPI::TEST(SOAuthorizationRedirect, InterceptionSucceedWith307Simple)):
(TestWebKitAPI::TEST(SOAuthorizationRedirect, InterceptionSucceedWith307CrossOrigin)):
(TestWebKitAPI::TEST(SOAuthorizationRedirect, InterceptionFailedWith307PUT)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/ServiceWorkerBasic.mm:
((ServiceWorker, FocusNotYetLoadedClient)):
((ServiceWorker, ServiceWorkerProcessSwapWithNoDelay)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST(SiteIsolation, LoadingCallbacksAndPostMessage)):
(TestWebKitAPI::TEST(SiteIsolation, PostMessageWithNotAllowedTargetOrigin)):
(TestWebKitAPI::TEST(SiteIsolation, ApplicationNameForUserAgent)):
(TestWebKitAPI::TEST(SiteIsolation, WebsitePoliciesCustomUserAgent)):
(TestWebKitAPI::TEST(SiteIsolation, WebsitePoliciesCustomUserAgentDuringCrossSiteProvisionalNavigation)):
(TestWebKitAPI::TEST(SiteIsolation, WebsitePoliciesCustomUserAgentDuringSameSiteProvisionalNavigation)):
(TestWebKitAPI::TEST(SiteIsolation, SandboxFlagsDuringNavigation)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKContentExtensionStore.mm:
(TEST_F(WKContentRuleListStoreTest, CrossOriginCookieBlocking)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKHTTPCookieStore.mm:
(TEST(WKHTTPCookieStore, WebSocketCookies)):
(TEST(WKHTTPCookieStore, WebSocketCookiesFromRedirect)):
(TEST(WKHTTPCookieStore, WebSocketCookiesThroughRedirect)):
(TEST(WKHTTPCookieStore, WebSocketSetCookiesThroughFirstPartyRedirect)):
(TEST(WKHTTPCookieStore, WebSocketSetCookiesThroughRedirectToThirdParty)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebExtensionAPIWebNavigation.mm:
(TestWebKitAPI::TEST(WKWebExtensionAPIWebNavigation, ErrorOccurredEventDuringLoad)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebViewEvaluateJavaScript.mm:
(TEST(WebKit, EvaluateJavaScriptInAttachments)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WebTransport.mm:
(TestWebKitAPI::TEST(WebTransport, DISABLED_ClientBidirectional)):
(TestWebKitAPI::TEST(WebTransport, DISABLED_Datagram)):
(TestWebKitAPI::TEST(WebTransport, DISABLED_Unidirectional)):
(TestWebKitAPI::TEST(WebTransport, DISABLED_ServerBidirectional)):
(TestWebKitAPI::TEST(WebTransport, DISABLED_NetworkProcessCrash)):
(TestWebKitAPI::TEST(WebTransport, DISABLED_Worker)):
* Tools/TestWebKitAPI/WebTransportServer.h:
* Tools/TestWebKitAPI/WebTransportServer.mm:
(TestWebKitAPI::WebTransportServer::Data::create):
(TestWebKitAPI::WebTransportServer::Data::Data):
(TestWebKitAPI::WebTransportServer::WebTransportServer):
* Tools/TestWebKitAPI/cocoa/HTTPServer.h:
* Tools/TestWebKitAPI/cocoa/HTTPServer.mm:
(TestWebKitAPI::HTTPServer::HTTPServer):

Canonical link: https://commits.webkit.org/291972@main
