+++ Tools/TestWebKitAPI/Tests/WebKitCocoa/WebTransport.mm
@@ -63,7 +63,7 @@ static void validateChallenge(NSURLAuthenticationChallenge *challenge, uint16_t 
 
 TEST(WebTransport, ClientBidirectional)
 {
-    WebTransportServer echoServer([](ConnectionGroup group) -> Task {
+    WebTransportServer echoServer([](ConnectionGroup group) -> ConnectionTask {
         auto connection = co_await group.receiveIncomingConnection();
         auto request = co_await connection.awaitableReceiveBytes();
         co_await connection.awaitableSend(WTFMove(request));
@@ -107,7 +107,7 @@ TEST(WebTransport, ClientBidirectional)
 
 TEST(WebTransport, Datagram)
 {
-    WebTransportServer echoServer([](ConnectionGroup group) -> Task {
+    WebTransportServer echoServer([](ConnectionGroup group) -> ConnectionTask {
         auto datagramConnection = group.createWebTransportConnection(ConnectionGroup::ConnectionType::Datagram);
         auto request = co_await datagramConnection.awaitableReceiveBytes();
         co_await datagramConnection.awaitableSend(WTFMove(request));
@@ -150,7 +150,7 @@ TEST(WebTransport, Datagram)
 
 TEST(WebTransport, Unidirectional)
 {
-    WebTransportServer echoServer([](ConnectionGroup group) -> Task {
+    WebTransportServer echoServer([](ConnectionGroup group) -> ConnectionTask {
         auto connection = co_await group.receiveIncomingConnection();
         auto request = co_await connection.awaitableReceiveBytes();
         auto serverUnidirectionalStream = group.createWebTransportConnection(ConnectionGroup::ConnectionType::Unidirectional);
@@ -198,7 +198,7 @@ TEST(WebTransport, Unidirectional)
 // FIXME: Fix and enable this test.
 TEST(WebTransport, DISABLED_ServerBidirectional)
 {
-    WebTransportServer echoServer([](ConnectionGroup group) -> Task {
+    WebTransportServer echoServer([](ConnectionGroup group) -> ConnectionTask {
         auto connection = co_await group.receiveIncomingConnection();
         auto request = co_await connection.awaitableReceiveBytes();
         auto serverBidirectionalStream = group.createWebTransportConnection(ConnectionGroup::ConnectionType::Bidirectional);
@@ -247,7 +247,7 @@ TEST(WebTransport, DISABLED_ServerBidirectional)
 
 TEST(WebTransport, NetworkProcessCrash)
 {
-    WebTransportServer echoServer([](ConnectionGroup group) -> Task {
+    WebTransportServer echoServer([](ConnectionGroup group) -> ConnectionTask {
         auto datagramConnection = group.createWebTransportConnection(ConnectionGroup::ConnectionType::Datagram);
         co_await datagramConnection.awaitableSend(@"abc");
         auto bidiConnection = group.createWebTransportConnection(ConnectionGroup::ConnectionType::Bidirectional);
@@ -419,7 +419,7 @@ TEST(WebTransport, NetworkProcessCrash)
 
 TEST(WebTransport, Worker)
 {
-    WebTransportServer transportServer([](ConnectionGroup group) -> Task {
+    WebTransportServer transportServer([](ConnectionGroup group) -> ConnectionTask {
         auto connection = co_await group.receiveIncomingConnection();
         auto request = co_await connection.awaitableReceiveBytes();
         auto serverBidirectionalStream = group.createWebTransportConnection(ConnectionGroup::ConnectionType::Bidirectional);
