+++ Tools/TestWebKitAPI/cocoa/HTTPServer.h
@@ -50,7 +50,7 @@ public:
     HTTPServer(std::initializer_list<std::pair<String, HTTPResponse>>, Protocol = Protocol::Http, CertificateVerifier&& = nullptr, SecIdentityRef = nullptr, std::optional<uint16_t> port = { });
     HTTPServer(Function<void(Connection)>&&, Protocol = Protocol::Http);
     enum class UseCoroutines : bool { Yes };
-    HTTPServer(UseCoroutines, Function<Task(Connection)>&&, Protocol = Protocol::Http);
+    HTTPServer(UseCoroutines, Function<ConnectionTask(Connection)>&&, Protocol = Protocol::Http);
     ~HTTPServer();
     uint16_t port() const;
     String origin() const;
